% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIFVMAR.R
\name{benchmarks}
\alias{benchmarks}
\title{Benchmark R and Rcpp Bootstraps.}
\description{
Use R package \code{microbenchmark} to compare the performance of Bootstrap and Rayleigh Random number generating in R and Rcpp.
}
\examples{
\dontrun{
data <- rnorm(100)
tm1 <- microbenchmark::microbenchmark(
  R = bootsR(data, 1000),
  C = boots(data, 1000)
)
print(summary(tm1)[,c(1,3,5,6)])

tm2 <- microbenchmark::microbenchmark(
  R = RayleighR(5e3,1,2e3),
  C = Rayleigh(5e3,1,2e3)
)
print(summary(tm2)[,c(1,3,5,6)])
}
}
