% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BootstrapC}
\alias{BootstrapC}
\alias{bootsC}
\title{Bootstrap in Rcpp}
\usage{
bootsC(x, B)
}
\arguments{
\item{x}{Data}

\item{B}{the sampling times}
}
\value{
a list of Bootstrap estimator
}
\description{
The bootstrap in Rcpp which can reduce the computation time largely when comparing with R.
}
\examples{
\dontrun{
data <- rnorm(100)
bootstrap_results <- bootsC(data, 1000)
}
}
