% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIFVMAR.R
\name{BootstrapR}
\alias{BootstrapR}
\alias{bootsR}
\title{Bootstrap in R}
\usage{
bootsR(x, B)
}
\arguments{
\item{x}{data}

\item{B}{sampling time}
}
\value{
a list of Bootstrap estimator
}
\description{
Bootstrap in R used for comparing the computation time with Rcpp.
}
\examples{
\dontrun{
data=rnorm(100)
bootsR(data,1000)
}
}
