% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIFVMAR.R
\name{MCMCMHR}
\alias{MCMCMHR}
\alias{RayleighR}
\title{Metropolis-Hastings sampler generating random number of Rayleigh distribution in R}
\usage{
RayleighR(m, sigma, b)
}
\arguments{
\item{m}{sample size}

\item{sigma}{the parameter of Rayleigh distribution}

\item{b}{burn-in size}
}
\value{
a list of Bootstrap estimator
}
\description{
Metropolis-Hastings sampler generates random number of Rayleigh distribution in R used for comparing the computation time with Rcpp.
}
\examples{
\dontrun{
RayleighR(m=5e3,sigma=1,b=2e3)
}
}
