% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RayleighC}
\alias{RayleighC}
\alias{Rayleigh}
\title{Rayleigh distribution random number}
\usage{
Rayleigh(m, sigma, b)
}
\arguments{
\item{m}{random numbers size}

\item{sigma}{Rayleigh distribution's parameter}

\item{b}{burn-in size}
}
\value{
a random sample from Rayleigh distribution
}
\description{
Generate random numbers from the Rayleigh distribution in Rcpp which can reduce the computation time largely when comparing with R.
}
\examples{
\dontrun{
RayRandom=Rayleigh(m=2000,sigma=4,b=100)
}
}
